Moved all the database things to it's own package
