+++ backend/database/conditions.go
@@ -1,4 +1,4 @@
-package main
+package database
 
 import (
 	"fmt"
